+++ kf_lib/fighting/fight/_base_fight.py
@@ -1,7 +1,7 @@
 import random
 
 from kf_lib.actors.names import GROUP_NAMES
-from kf_lib.constants.experience import BASE_FIGHT_EXP, LOSER_EXP
+from kf_lib.constants import experience
 from kf_lib.ui import cls, menu, pak
 
 
@@ -194,7 +194,7 @@ class BaseFight(object):
         if not self.winners:
             # draw: everyone gets a fixed, reduced amount of exp
             for p in self.players:
-                p.gain_exp(round(BASE_FIGHT_EXP / DRAW_EXP_DIVISOR))
+                p.gain_exp(round(experience.BASE_FIGHT_EXP / DRAW_EXP_DIVISOR))
             self.handle_player_stats()
             self.main_player.pak()
             return
@@ -203,8 +203,8 @@ class BaseFight(object):
         winners_yield = self.aggregate_exp_yield(self.winners)
         losers_yield = self.aggregate_exp_yield(self.losers)
         winners_diff = (losers_yield / winners_yield) ** 1.5
-        winners_gain = winners_diff * BASE_FIGHT_EXP
-        losers_gain = LOSER_EXP
+        winners_gain = winners_diff * experience.BASE_FIGHT_EXP
+        losers_gain = experience.LOSER_EXP
         for p in self.players:
             if p in self.winners:
                 exp = self.handle_exp_bonuses(p, winners_gain)